Add template specialization for base class
The method binding is resolved at compile time. Therefore, the
compiler looks for IEncodersRaw:: instead of IEncodersTimedRaw::
when binding to a method that relates to the former, not accounting
for the inheritance chain (resolved on runtime due to *late* binding).
So, the linker complained on runtime due to not finding the proper
definition of, say, IEncodersTimedRaw::setEncoderRaw.